// front/src/types/api/api-types.ts
export interface BaseModel {
  id: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string | null;
}
// ...
export interface APIAdmissionInfo extends BaseModel {
  major_id: number;
  enrollment: number;
  academic_year: number;
  status: string;
}

export interface APITestType extends BaseModel {
  admission_schedule_id: number;
  name: string;
  subjects: APISubject[];
}

export interface APISubject extends BaseModel {
  test_type_id: number;
  name: string;
  score: number;
  percentage: number;
  display_order: number;
}